Rebrand Datasets2 with new path layout as Datasets3
