+++ kart/repo.py
@@ -19,12 +19,12 @@ from .exceptions import (
     InvalidOperation,
     NotFound,
     NO_REPOSITORY,
-    UNSUPPORTED_VERSION,
 )
 from .repo_version import (
-    SUPPORTED_REPO_VERSION,
-    SUPPORTED_DATASET_CLASS,
+    DEFAULT_NEW_REPO_VERSION,
+    ensure_supported_repo_version,
     get_repo_version,
+    dataset_class_for_version,
 )
 from .structure import RepoStructure
 from .timestamps import tz_offset_to_minutes
@@ -337,7 +337,7 @@ class KartRepo(pygit2.Repository):
         version_key = KartConfigKeys.BRANDED_REPOSTRUCTURE_VERSION_KEYS[
             self.BRANDING_FOR_NEW_REPOS
         ]
-        self.config[version_key] = str(SUPPORTED_REPO_VERSION)
+        self.config[version_key] = str(DEFAULT_NEW_REPO_VERSION)
 
         # Bare-style Kart repos are always implemented as bare git repos:
         if self.is_bare_style:
@@ -351,18 +351,7 @@ class KartRepo(pygit2.Repository):
         BaseWorkingCopy.write_config(self, wc_location, bare)
 
     def ensure_supported_version(self):
-        from .cli import get_version
-
-        if self.version != SUPPORTED_REPO_VERSION:
-            message = (
-                f"This Kart repo uses Datasets v{self.version}, "
-                f"but Kart {get_version()} only supports Datasets v{SUPPORTED_REPO_VERSION}.\n"
-            )
-            if self.version < SUPPORTED_REPO_VERSION:
-                message += "Use `kart upgrade SOURCE DEST` to upgrade this repo to the supported version."
-            else:
-                message += "Get the latest version of Kart to work with this repo."
-            raise InvalidOperation(message, exit_code=UNSUPPORTED_VERSION)
+        ensure_supported_repo_version(self.version)
 
     def write_readme(self):
         readme_filename = f"{self.branding.upper()}_README.txt"
@@ -511,7 +500,7 @@ class KartRepo(pygit2.Repository):
     def structure(self, refish="HEAD"):
         """Get the structure of this Kart repository at a particular revision."""
         self.ensure_supported_version()
-        return RepoStructure(self, refish, dataset_class=SUPPORTED_DATASET_CLASS)
+        return RepoStructure(self, refish, dataset_class=self.dataset_class)
 
     def datasets(self, refish="HEAD"):
         """
@@ -519,6 +508,11 @@ class KartRepo(pygit2.Repository):
         Equivalent to: self.structure(refish).datasets
         """
         return self.structure(refish).datasets
+
+    @property
+    def dataset_class(self):
+        self.ensure_supported_version()
+        return dataset_class_for_version(self.version)
 
     @property
     def working_copy(self):
