+++ tests/data/upgrade/old-bare.tgz
