+++ tests/test_structure.py
@@ -10,8 +10,8 @@ import pygit2
 import pytest
 
 from kart import init, fast_import
-from kart.dataset2 import Dataset2
-from kart.dataset2_paths import IntPathEncoder, MsgpackHashPathEncoder
+from kart.dataset3 import Dataset3
+from kart.dataset3_paths import IntPathEncoder, MsgpackHashPathEncoder
 from kart.exceptions import INVALID_OPERATION
 from kart.sqlalchemy.gpkg import Db_GPKG
 from kart.schema import Schema
@@ -54,7 +54,7 @@ GPKG_IMPORTS = (
 def _import_check(repo_path, table, source_gpkg):
     repo = KartRepo(repo_path)
     dataset = repo.datasets()[table]
-    assert dataset.VERSION == 2
+    assert dataset.VERSION == 3
 
     with Db_GPKG.create_engine(source_gpkg).connect() as conn:
         num_rows = conn.execute(f"SELECT COUNT(*) FROM {table};").fetchone()[0]
@@ -62,10 +62,10 @@ def _import_check(repo_path, table, source_gpkg):
     o = subprocess.check_output(["git", "ls-tree", "-r", "-t", "HEAD", table])
     print("\n".join(l.decode("utf8") for l in o.splitlines()[:20]))
 
-    if dataset.VERSION != 2:
+    if dataset.VERSION != 3:
         raise NotImplementedError(dataset.VERSION)
 
-    re_paths = r"^\d{6} blob [0-9a-f]{40}\t%s/.sno-dataset/feature/.*$" % table
+    re_paths = r"^\d{6} blob [0-9a-f]{40}\t%s/.table-dataset/feature/.*$" % table
     git_paths = [m for m in re.findall(re_paths, o.decode("utf-8"), re.MULTILINE)]
     assert len(git_paths) == num_rows
 
@@ -868,7 +868,7 @@ def test_pk_encoder_legacy_hashed(data_archive_readonly):
 
 
 def test_pk_encoder_string_pk():
-    ds = Dataset2(None, "mytable")
+    ds = Dataset3(None, "mytable")
     schema = Schema.from_column_dicts(
         [{"name": "mypk", "dataType": "text", "id": "abc123"}]
     )
@@ -878,14 +878,15 @@ def test_pk_encoder_string_pk():
     assert e.encoding == "base64"
     assert e.branches == 64
     assert e.levels == 4
-    assert ds.encode_1pk_to_path("") == "mytable/.sno-dataset/feature/I/6/M/_/kaA="
+    assert ds.encode_1pk_to_path("") == "mytable/.table-dataset/feature/I/6/M/_/kaA="
     assert (
-        ds.encode_1pk_to_path("Dave") == "mytable/.sno-dataset/feature/s/v/7/j/kaREYXZl"
+        ds.encode_1pk_to_path("Dave")
+        == "mytable/.table-dataset/feature/s/v/7/j/kaREYXZl"
     )
 
 
 def test_pk_encoder_int_pk():
-    ds = Dataset2(None, "mytable")
+    ds = Dataset3(None, "mytable")
     schema = Schema.from_column_dicts(
         [
             {
@@ -909,20 +910,20 @@ def test_pk_encoder_int_pk():
     with pytest.raises(TypeError):
         ds.encode_1pk_to_path(0.1)
 
-    assert ds.encode_1pk_to_path(0) == "mytable/.sno-dataset/feature/A/A/A/A/kQA="
-    assert ds.encode_1pk_to_path(1) == "mytable/.sno-dataset/feature/A/A/A/A/kQE="
-    assert ds.encode_1pk_to_path(-1) == "mytable/.sno-dataset/feature/_/_/_/_/kf8="
+    assert ds.encode_1pk_to_path(0) == "mytable/.table-dataset/feature/A/A/A/A/kQA="
+    assert ds.encode_1pk_to_path(1) == "mytable/.table-dataset/feature/A/A/A/A/kQE="
+    assert ds.encode_1pk_to_path(-1) == "mytable/.table-dataset/feature/_/_/_/_/kf8="
     assert (
-        ds.encode_1pk_to_path(1181) == "mytable/.sno-dataset/feature/A/A/A/S/kc0EnQ=="
+        ds.encode_1pk_to_path(1181) == "mytable/.table-dataset/feature/A/A/A/S/kc0EnQ=="
     )
     # trees hit wraparound with large PKs, but don't break
     assert (
         ds.encode_1pk_to_path(64 ** 5)
-        == "mytable/.sno-dataset/feature/A/A/A/A/kc5AAAAA"
+        == "mytable/.table-dataset/feature/A/A/A/A/kc5AAAAA"
     )
     assert (
         ds.encode_1pk_to_path(-(64 ** 5))
-        == "mytable/.sno-dataset/feature/A/A/A/A/kdLAAAAA"
+        == "mytable/.table-dataset/feature/A/A/A/A/kdLAAAAA"
     )
 
 
@@ -1081,7 +1082,7 @@ def test_write_feature_performance(
 
             source = ImportSource.open(data / source_gpkg, table=table)
             with source:
-                dataset = Dataset2(None, table)
+                dataset = Dataset3(None, table)
                 feature_iter = itertools.cycle(list(source.features()))
 
                 index = pygit2.Index()
@@ -1127,7 +1128,7 @@ def test_fast_import(data_archive, tmp_path, cli_runner, chdir):
             assert repo.head.shorthand == "main"
 
             dataset = repo.datasets()[table]
-            assert dataset.VERSION == 2
+            assert dataset.VERSION == 3
 
             # has a single commit
             assert len([c for c in repo.walk(repo.head.target)]) == 1
